feat: implemented logout api

+++ packages/louder-server/src/api/models/AuthToken.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-underscore-dangle */
 import { prisma } from "../database/prisma";
 
 export type AuthTokenSerialized = {
@@ -35,7 +36,27 @@ export default class AuthToken {
 
   public static async getByUserId(userId: string) {
     // eslint-disable-next-line no-return-await
-    return await prisma.authToken.findUnique({
+    const _authToken = await prisma.authToken.findUnique({
+      where: {
+        user_id: userId,
+      },
+    });
+
+    if (!_authToken) return null;
+
+    const authToken = new AuthToken({
+      auth_token: _authToken.auth_token,
+      refresh_token: _authToken.refresh_token,
+      user_id: _authToken.user_id,
+      id: _authToken.id,
+      created_at: _authToken.created_at,
+    });
+
+    return authToken;
+  }
+
+  public static async deleteByUserId(userId: string) {
+    await prisma.authToken.delete({
       where: {
         user_id: userId,
       },
@@ -52,17 +73,23 @@ export default class AuthToken {
 
   public async create() {
     const existing = await AuthToken.getByUserId(this.user_id);
-    if (existing) {
+    if (existing && existing.id) {
       await AuthToken.deleteById(existing.id);
     }
     // eslint-disable-next-line no-return-await
-    return await prisma.authToken.create({
+    const _authToken = await prisma.authToken.create({
       data: {
         auth_token: this.auth_token,
         refresh_token: this.refresh_token,
         user_id: this.user_id,
       },
     });
+
+    const authToken = new AuthToken({
+      ..._authToken,
+    });
+
+    return authToken;
   }
 
   public serialize(): AuthTokenSerialized {
